+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- No more Gradle 8 warnings about `without the corresponding declaration via 'Task#usesService'`. ([#47](https://github.com/diffplug/spotless-changelog/issues/47))
 
 ## [3.0.1] - 2023-01-30
 ### Fixed